move new visualization functions to qc
