+++ docs/img/documentation_images/qc_color_chip_comparison/output.png
