+++ docs/img/documentation_images/qc_color_correction_scatter/am003_sv_ex.png
