+++ mkdocs.yml
@@ -144,6 +144,8 @@ nav:
       - 'Print Image': print_image.md
       - 'Plot Image': plot_image.md
       - "Quality Control Tools":
+        - 'Color Correction Scatter Plot': qc_color_correction_scatter.md
+        - 'Color Chip Comparison Plot': qc_color_chip_comparison.md
         - "Exposure": qc_exposure.md
         - "Quick Color Check": quick_color_check.md
       - 'Read Image': read_image.md
@@ -208,8 +210,6 @@ nav:
       - 'Visualization Methods':
           - 'Auto Threshold Methods': visualize_auto_threshold_methods.md
           - 'Chlorophyll Fluorescence': visualize_chlorophyll_fluorescence.md
-          - 'Color Correction Scatter Plot': visualize_color_correction_scatter.md
-          - 'Color Chip Comparison Plot': visualize_color_chip_comparison.md
           - 'Colorize Label Image': visualize_colorize_label_img.md
           - 'Colorize Masks': visualize_colorize_masks.md
           - 'Colorspaces': visualize_colorspace.md
