+++ plantcv/plantcv/qc/__init__.py
@@ -1,5 +1,7 @@
 from plantcv.plantcv.qc.exposure import exposure
 from plantcv.plantcv.qc.quick_color_check import quick_color_check
+from plantcv.plantcv.qc.color_correction_scatter import color_correction_plot
+from plantcv.plantcv.qc.color_chip_comparison import color_chip_comparison
 
 
-__all__ = ["exposure", "quick_color_check"]
+__all__ = ["exposure", "quick_color_check", "color_correction_plot", "color_chip_comparison"]
