+++ plantcv/plantcv/qc/color_chip_comparison.py
